@@ -1,26 +1,41 @@
 [tool.poetry]
 name = "pybotx-smartapp-smart-logger"
-version = "0.2.6"
+version = "0.3.0"
 description = ""
 authors = ["Arseniy Zhiltsov <arseniy.zhiltsov@ccsteam.ru>"]
 
 [tool.poetry.dependencies]
-python = ">=3.8, <3.11"
+python = ">=3.10,<3.14"
 
-pybotx = ">=0.42.0,<0.45.0"
-pybotx-smart-logger = ">=0.6.7,<0.8.0"
-pybotx-smartapp-rpc = ">=0.5.0,<0.6.0"
+pybotx = ">=0.75.0,<1.0.0"
+pybotx-smart-logger = "^0.12.0"
+pybotx-smartapp-rpc = "^0.13.0"
 
-[tool.poetry.dev-dependencies]
-black = "22.3.0"
-isort = "^4.3"
-autoflake = "^1.4"
-mypy = "0.910"
-wemake-python-styleguide = "0.16.1"
+[tool.poetry.group.dev.dependencies]
+mypy = "^1.16.1"
+pytest = "^8.4.1"
+pytest-asyncio = "^1.1.0"
+pytest-cov = "^6.2.1"
+pytest-mock = "^3.15.1"
+pytest-xdist = "^3.8.0"
+factory-boy = "^3.3.3"
+faker = "^37.8.0"
+freezegun = "^1.5.5"
+ruff = "^0.12.0"
 
-pytest = "^7.0.1"
-pytest-asyncio = "^0.18.1"
-pytest-cov = "^3.0.0"
+[tool.ruff]
+target-version = "py310"
+line-length = 88
+
+[tool.ruff.lint]
+select = ["E", "F", "I", "UP", "W", "B"]
+
+[tool.ruff.lint.per-file-ignores]
+"pybotx_smartapp_smart_logger/__init__.py" = ["F401"]
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
