@@ -1,5 +1,9 @@
+from pybotx_smartapp_smart_logger.smartapp_event_wrapper import wrap_system_event
 from pybotx_smartapp_smart_logger.smartapp_exception_handler import (
     smartapp_exception_handler,
 )
 
-__all__ = ("smartapp_exception_handler",)
+__all__ = (
+    "smartapp_exception_handler",
+    "wrap_system_event",
+)
